refactor: public pages and split qr page into data and page
