+++ src/google/mod.rs
@@ -1,2 +1,3 @@
 pub mod auth;
+pub mod http_client;
 pub mod sheets;
