+++ src/pages/qr/mod.rs
@@ -1,0 +1,2 @@
+pub mod data;
+pub mod page;
